data_ingestion/requirements.txt
requests 
pyyaml